requirements.txt
aiogram==3.0.0
SQLAlchemy==2.0.35
aiohttp==3.9.4
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqladmin==0.15.0
